Inline som JS code into the module
There are some functions, such as dispatch_function_in_event_loop, that
require JS code to be run before being available.  In order to
bullet-proof this dependency, we just include that JS code into the C
code that we compile, and make sure it is always run when creating an
instance of duktape.