Properly cleanup duktape stack
When dealing with global objects and properties, we were not always
popping elements from the duktape stack, especially in failure cases.
This can be found out by examining the duktape stack size, which kept
growing forever in some cases.  Fix that and add comments.

Also added a test (which aborts when failing, but what are you going to
do) that now passes.