Only create seen hash once
We were calling newHV() and later hv_undef() every time we converted a
Perl data structure into its JS counterpart (and viceversa); this hash
keeps track of items we have already converted, to properly support
cyclic references.  Alas, it seems hv_undef() does not actually delete
the hash, so we had a memory leak.  We now keep the seen hash around and
clear it every time we use it, which is also potentially more efficient.