Handle missing elements in a returned array.
Issue #30: Missing JS array items cause Duktape’s duk_get_prop_index()
to return false. In that case pl_duk_to_perl_impl() was neglecting to
duk_pop(ctx), which caused the next read on the array to be an attempt
to read index 1 from undefined, which triggered an abort().

This fixes that by ensuring that we always duk_pop(ctx) in the relevant
piece of code and return Perl undef to represent missing array items.